updated css for web view
