+++ src/collections/projects/sample/index.mdx
@@ -1,7 +1,7 @@
 ---
 title: "sample"                
 thumbnail: ./layer5.svg          
-published: false
+published: true
 ---
 
 import { ProjectWrapper } from "../Project.style.js";
